// src/app/people/interfaces/person.interface.ts
export interface person {
        imageUrl: {
            w200: string;
            w400: string;
            w1080: string;
            w1920: string;
            w2560: string;
        },
        block: {
            title: string,
            description: string,
            link: string,
            text: string
        }
}